+++ packages/core/src/tools/utils.ts
@@ -1,3 +1,5 @@
+import type { TimeoutId } from '../browser/timer'
+import { setTimeout, clearTimeout } from '../browser/timer'
 import { display } from './display'
 import { monitor } from './monitor'
 
@@ -525,15 +527,16 @@ export function combine(...sources: any[]): unknown {
   return destination as unknown
 }
 
-export type TimeoutId = ReturnType<typeof setTimeout>
-
+/**
+ * Use 'requestIdleCallback' when available: it will throttle the mutation processing if the
+ * browser is busy rendering frames (ex: when frames are below 60fps). When not available, the
+ * fallback on 'requestAnimationFrame' will still ensure the mutations are processed after any
+ * browser rendering process (Layout, Recalculate Style, etc.), so we can serialize DOM nodes efficiently.
+ *
+ * Note: check both 'requestIdleCallback' and 'cancelIdleCallback' existence because some polyfills only implement 'requestIdleCallback'.
+ */
 export function requestIdleCallback(callback: () => void, opts?: { timeout?: number }) {
-  // Use 'requestIdleCallback' when available: it will throttle the mutation processing if the
-  // browser is busy rendering frames (ex: when frames are below 60fps). When not available, the
-  // fallback on 'requestAnimationFrame' will still ensure the mutations are processed after any
-  // browser rendering process (Layout, Recalculate Style, etc.), so we can serialize DOM nodes
-  // efficiently.
-  if (window.requestIdleCallback) {
+  if (window.requestIdleCallback && window.cancelIdleCallback) {
     const id = window.requestIdleCallback(monitor(callback), opts)
     return () => window.cancelIdleCallback(id)
   }