+++ packages/rum/src/domain/record/observers/index.ts
@@ -1,0 +1,3 @@
+export { initObservers } from './observers'
+export { InputCallback, initInputObserver } from './inputObserver'
+export { initMutationObserver, MutationCallBack, RumMutationRecord } from './mutationObserver'