+++ packages/rum/src/domain/record/utils.ts
@@ -6,13 +6,6 @@ export function isTouchEvent(event: MouseEvent | TouchEvent): event is TouchEven
   return Boolean((event as TouchEvent).changedTouches)
 }
 
-export function forEach<List extends { [index: number]: any }>(
-  list: List,
-  callback: (value: List[number], index: number, parent: List) => void
-) {
-  Array.prototype.forEach.call(list, callback as any)
-}
-
 export function assembleIncrementalSnapshot<Data extends BrowserIncrementalData>(
   source: Data['source'],
   data: Omit<Data, 'source'>